refactor select components
