+++ app/src/components/VerticalDivider.tsx
@@ -1,0 +1,7 @@
+import { FC } from 'react'
+
+const VerticalDivider: FC = () => (
+  <div className="ml-2 h-[1.625rem] border-1 border-turtle-level3" />
+)
+
+export default VerticalDivider
